refactor : 계산기와 자동차 경주 게임의 메인 함수 패키지 별로 분리 / 실행 결과 출력 추가 / 자동차 경주 게임의 우승자 이름을 만드는 기능을 메소드로 추출
