+++ src/main/java/application/racing/Output.java
@@ -8,7 +8,7 @@ public class Output {
     }
 
     public static void printRacingResultMessage() {
-        System.out.println("실행 결과");
+        System.out.println("\n실행 결과");
     }
 
     public void printPositionDuringRacing(List<Car> cars) {
@@ -27,6 +27,11 @@ public class Output {
     }
 
     public static void printWinner(List<String> winners) {
-        System.out.println(String.join(", ", winners) + "가 최종 우승했습니다.");
+        String winnerNameString = makeWinnerNameString(winners);
+        System.out.println(winnerNameString + "가 최종 우승했습니다.");
+    }
+
+    private static String makeWinnerNameString(List<String> winners) {
+        return String.join(", ", winners);
     }
 }
